new property tweaks
